refactor: domain 패키지 내 Model 클래스 불필요 변수, 함수 제거
