--- src/main/java/vendingmachine/controller/CoinController.java
@@ -1,5 +1,0 @@
-package vendingmachine.controller;
-
-public class CoinController {
-
-}
