+++ src/main/java/vendingmachine/domain/Product.java
@@ -10,6 +10,7 @@ public class Product {
         this.price = price;
         this.amount = amount;
     }
+
     public void sell() {
         amount--;
     }
@@ -17,9 +18,7 @@ public class Product {
     public boolean existStock() {
         return this.amount > 0;
     }
-    public int getAmount(){
-        return amount;
-    }
+
     public int isLessThanMinimumPrice(int minimumPrice) {
         if (minimumPrice > this.price) {
             return this.price;
@@ -33,9 +32,5 @@ public class Product {
 
     public String getName() {
         return this.name;
-    }
-    // After Test
-    public String toString(){
-        return name+" : "+price+"원, "+amount+"개";
     }
 }
